dev_requirements.txt
mock
coverage
pygments
nose
pycodestyle
wheel
pylint
pytest
twine
jinja2
python-dateutil
